+++ Runtime/Services/ServiceLocatorBase.cs
@@ -1,11 +1,11 @@
 using Lunar.Interfaces;
 
-namespace Lunar.Adapters.Unity
+namespace Lunar.Adapters.Unity.Services
 {
     public interface IServiceLocator
     {
         IResources Resources { get; }
-        //IResourcesAsync ResourcesAsync { get; }
+        IResourcesAsync ResourcesAsync { get; }
         IInputActions InputActions { get; }
         ILogger Debug { get; }
     }
