+++ Runtime/Systems/GameObjectSyncSystem.cs
@@ -1,5 +1,6 @@
 using System;
 using Arch.Core;
+using Lunar.Adapters.Unity.Utils;
 using Lunar.Components;
 using Lunar.Systems;
 using UnityEngine;
